back-end-bus-ticket-service/staff-dispatch-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
sqlalchemy==2.0.29
asyncpg==0.29.0
aioredis==2.0.1
httpx==0.27.0
PyJWT==2.8.0
redis==5.0.1
requests==2.31.0
python-dotenv
supabase

